NinchatMediaActivity refactoring  - Move NinchatMediaActivity to it's own package
