+++ ninchatsdk/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
             android:name=".ninchatreview.view.NinchatReviewActivity"
             android:theme="@style/NinchatTheme" />
         <activity
-            android:name=".activities.NinchatMediaActivity"
+            android:name=".ninchatmedia.view.NinchatMediaActivity"
             android:theme="@style/NinchatTheme" />
 
         <activity
